Update test, pass environment variable HYPOLAS_HEALTHCHECK_ID

+++ hypolashlckhelpers.go
@@ -25,11 +25,11 @@ func resolveCMD(cmdString string) string {
 	inputString := strings.TrimSpace(cmdString)
 	strCommand := getStringInBetween(inputString, "#CMDSTART#", "#CMDEND#")
 	stringToReplace := "#CMDSTART#" + strCommand + "#CMDEND#"
-	logf.VarDebug(strCommand, "strCommand")
-	logf.VarDebug(stringToReplace, "stringToReplace")
+	log.VarDebug(strCommand, "strCommand")
+	log.VarDebug(stringToReplace, "stringToReplace")
 
 	cmdArgs := strings.Split(strings.TrimSpace(strCommand), " ")
-	logf.VarDebug(cmdArgs, "cmdArgs")
+	log.VarDebug(cmdArgs, "cmdArgs")
 
 	var cmd *exec.Cmd
 	if len(cmdArgs) == 1 {
@@ -42,7 +42,7 @@ func resolveCMD(cmdString string) string {
 	cmd.Env = os.Environ()
 	err := cmd.Run()
 	if err != nil {
-		logf.Err.Fatalf("%s\n", err)
+		log.Err.Fatalf("%s\n", err)
 	}
 
 	/*
@@ -51,10 +51,10 @@ func resolveCMD(cmdString string) string {
 	cmdResult := strings.TrimFunc(stdout.String(), func(r rune) bool {
 		return !unicode.IsGraphic(r)
 	})
-	logf.VarDebug(cmdResult, "cmdResult")
+	log.VarDebug(cmdResult, "cmdResult")
 
 	detectedCMDValue := strings.Replace(inputString, stringToReplace, cmdResult, -1)
-	logf.VarDebug(detectedCMDValue, "detectedCMDValue")
+	log.VarDebug(detectedCMDValue, "detectedCMDValue")
 
 	return detectedCMDValue
 }
